Update PR number and rename files
