+++ packages/core/src/sagas.js
@@ -4,7 +4,7 @@ import clearSuggestedActionsOnPostActivitySaga from './sagas/clearSuggestedActio
 import connectionStatusUpdateSaga from './sagas/connectionStatusUpdateSaga';
 import connectSaga from './sagas/connectSaga';
 import detectSlowConnectionSaga from './sagas/detectSlowConnectionSaga';
-import emitTypingIndicatorSaga from './sagas/emitTypingIndicatorSaga';
+import emitTypingIndicatorToPostActivitySaga from './sagas/emitTypingIndicatorToPostActivitySaga';
 import incomingActivitySaga from './sagas/incomingActivitySaga';
 import markAllAsSpokenOnStopSpeakActivitySaga from './sagas/markAllAsSpokenOnStopSpeakActivitySaga';
 import postActivitySaga from './sagas/postActivitySaga';
@@ -27,7 +27,7 @@ export default function* sagas() {
   yield fork(connectionStatusUpdateSaga);
   yield fork(connectSaga);
   yield fork(detectSlowConnectionSaga);
-  yield fork(emitTypingIndicatorSaga);
+  yield fork(emitTypingIndicatorToPostActivitySaga);
   yield fork(incomingActivitySaga);
   yield fork(markAllAsSpokenOnStopSpeakActivitySaga);
   yield fork(postActivitySaga);
