+++ packages/core/src/sagas/emitTypingIndicatorToPostActivitySaga.js
